+++ app/Livewire/Admin/Role/Index.php
@@ -6,11 +6,13 @@ use Livewire\Component;
 use Spatie\Permission\Models\Role;
 use Livewire\WithPagination;
 use Livewire\Attributes\Layout;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+
 
 #[Layout('layouts.app')]
 class Index extends Component
 {
-    use WithPagination;
+    use WithPagination, AuthorizesRequests;
 
     public $showForm = false;
     public $editMode = false;
@@ -20,12 +22,15 @@ class Index extends Component
 
     public function create()
     {
+        $this->authorize('create_role');
         $this->reset(['roleId', 'editMode']);
         $this->showForm = true;
     }
 
     public function edit($id)
     {
+       
+        $this->authorize('edit_role');
         $this->roleId = $id;
         $this->editMode = true;
         $this->showForm = true;
@@ -33,6 +38,7 @@ class Index extends Component
 
     public function delete($id)
     {
+        $this->authorize('delete_role');
         Role::findOrFail($id)->delete();
     }
 
@@ -47,5 +53,9 @@ class Index extends Component
             'roles' => Role::latest()->paginate(10),
         ]);
     }
+
+     // $this->authorize('create_role');
+
+        // $this->authorize('edit_role');
 }
